code: index, glance, pdf
Custom non-CRUD routes are used for returning a more tabular view of upcoming operations, as the main index page will be rather bespoke. Later, I'm planning on adding a PDF export library to the software, and this route will allow the operation to be exported. In future, I'd consider allowing other data to be exportable too.